[modmail] extract notehandler to own class
